+++ common/src/main/java/com/kpsys/common/requests/ClientCustomDataRequest.java
@@ -7,5 +7,5 @@ import javax.validation.constraints.NotNull;
 @Data
 public class ClientCustomDataRequest {
     @NotNull
-    Long clientId;
+    Integer clientId;
 }
